Rename segment folder to segmentation
